Mostly cleaned up and ready to go
